requirements.dev.txt
-r requirements.txt
# Type hinting extras
django-types == 0.16.0
djangorestframework-types == 0.8.0
